all: Move most of graph::components::ethereum::network to the chain
