+++ core/Cargo.toml
@@ -10,6 +10,9 @@ bytes = "0.5"
 futures01 = { package="futures", version="0.1.31" }
 futures = { version="0.3.4", features=["compat"] }
 graph = { path = "../graph" }
+# This dependency is temporary. The multiblockchain refactoring is not
+# finished as long as this dependency exists
+graph-chain-ethereum = { path = "../chain/ethereum" }
 lazy_static = "1.2.0"
 lru_time_cache = "0.11"
 semver = "0.10.0"
