+++ runtime/wasm/Cargo.toml
@@ -13,6 +13,9 @@ graph = { path = "../../graph" }
 graph-graphql = { path = "../../graphql" }
 bs58 = "0.4.0"
 graph-runtime-derive = { path = "../derive" }
+# This dependency is temporary. The multiblockchain refactoring is not
+# finished as long as this dependency exists
+graph-chain-ethereum = { path = "../../chain/ethereum" }
 semver = "0.10.0"
 lazy_static = "1.4"
 uuid = { version = "0.8.1", features = ["v4"] }
